+++ output.txt
@@ -1,0 +1,11 @@
+Your flight departs from Hannover Airport, and your destination is Bremen Airport.
+
+1. 09 May 2022
+2. 07:18PM (-02:00)
+3. 02:54PM (+00:00)
+4. 03:30AM (+11:00)
+5. 03:09AM (+00:00)
+6. 04:08 (+13:00)
+7. 17:54 (+00:00)
+8. 15:29 (-11:00)
+9. 21:43 (+00:00)
